+++ packages/liveblocks-core/src/constants.ts
@@ -1,1 +1,3 @@
 export const DEFAULT_BASE_URL = "https://api.liveblocks.io";
+
+export const MENTION_CHARACTER = "@";