+++ packages/liveblocks-react-lexical/src/mentions/mention-plugin.tsx
@@ -9,7 +9,11 @@ import {
   useFloating,
 } from "@floating-ui/react-dom";
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import { assertNever, type MentionData } from "@liveblocks/core";
+import {
+  assertNever,
+  MENTION_CHARACTER,
+  type MentionData,
+} from "@liveblocks/core";
 import { useRoom } from "@liveblocks/react";
 import {
   useCreateTextMention,
@@ -17,7 +21,13 @@ import {
   useLayoutEffect,
   useMentionSuggestions,
 } from "@liveblocks/react/_private";
-import { Avatar, User } from "@liveblocks/react-ui/_private";
+import {
+  Avatar,
+  Group,
+  GroupDescription,
+  User,
+  UsersIcon,
+} from "@liveblocks/react-ui/_private";
 import type { EditorState, NodeKey, NodeMutation, TextNode } from "lexical";
 import {
   $createRangeSelection,
@@ -37,6 +47,11 @@ import { useCallback, useEffect, useState } from "react";
 import { createPortal } from "react-dom";
 
 import {
+  $createGroupMentionNode,
+  $isGroupMentionNode,
+  GroupMentionNode,
+} from "./group-mention-node";
+import {
   $createMentionNode,
   $isMentionNode,
   MentionNode,
@@ -48,13 +63,11 @@ import {
   SuggestionsContext,
 } from "./suggestions";
 
-const MENTION_TRIGGER = "@";
-
 const PUNCTUATIONS =
   "\\.,\\+\\*\\?\\$\\@\\|#{}\\(\\)\\^\\-\\[\\]\\\\/!%'\"~=<>_:;";
 
 // Characters we expect to see in a mention (non-space, non-punctuation).
-const VALID_CHARACTERS = "[^" + MENTION_TRIGGER + PUNCTUATIONS + "\\s]";
+const VALID_CHARACTERS = "[^" + MENTION_CHARACTER + PUNCTUATIONS + "\\s]";
 
 const VALID_JOINS =
   "(?:" +
@@ -70,7 +83,7 @@ const LENGTH_LIMIT = 75;
 const MentionRegex = new RegExp(
   "(^|\\s|\\()(" +
     "[" +
-    MENTION_TRIGGER +
+    MENTION_CHARACTER +
     "]" +
     "((?:" +
     VALID_CHARACTERS +
@@ -167,7 +180,7 @@ export function MentionPlugin() {
   const deleteTextMention = useDeleteTextMention();
 
   useEffect(() => {
-    function $handleMutation(
+    function $handleMentionMutation(
       mutations: Map<NodeKey, NodeMutation>,
       {
         prevEditorState,
@@ -181,22 +194,62 @@ export function MentionPlugin() {
             const node = $getNodeByKey(key);
             if (node === null) return;
 
-            if (!$isMentionNode(node)) return;
-            createTextMention(node.getUserId(), node.getId());
+            if ($isMentionNode(node)) {
+              createTextMention(node.getId(), {
+                kind: "user",
+                id: node.getUserId(),
+              });
+            }
           });
         } else if (mutation === "destroyed") {
           prevEditorState.read(() => {
             const node = $getNodeByKey(key);
             if (node === null) return;
 
-            if (!$isMentionNode(node)) return;
-            deleteTextMention(node.getId());
+            if ($isMentionNode(node)) {
+              deleteTextMention(node.getId());
+            }
           });
         }
       }
     }
 
-    return editor.registerMutationListener(
+    function $handleGroupMentionMutation(
+      mutations: Map<NodeKey, NodeMutation>,
+      {
+        prevEditorState,
+      }: {
+        prevEditorState: EditorState;
+      }
+    ) {
+      for (const [key, mutation] of mutations) {
+        if (mutation === "created") {
+          editor.getEditorState().read(() => {
+            const node = $getNodeByKey(key);
+            if (node === null) return;
+
+            if ($isGroupMentionNode(node)) {
+              createTextMention(node.getId(), {
+                kind: "group",
+                id: node.getGroupId(),
+                userIds: node.getUserIds(),
+              });
+            }
+          });
+        } else if (mutation === "destroyed") {
+          prevEditorState.read(() => {
+            const node = $getNodeByKey(key);
+            if (node === null) return;
+
+            if ($isGroupMentionNode(node)) {
+              deleteTextMention(node.getId());
+            }
+          });
+        }
+      }
+    }
+
+    const unsubscribeMentionMutationListener = editor.registerMutationListener(
       MentionNode,
       (mutations, payload) => {
         // Ignore mutations to MentionNode (creation/updates/deletions) that are caused by collaboration (remote users) or history merge.
@@ -207,9 +260,30 @@ export function MentionPlugin() {
           return;
         }
 
-        $handleMutation(mutations, payload);
+        $handleMentionMutation(mutations, payload);
       }
     );
+
+    const unsubscribeGroupMentionMutationListener =
+      editor.registerMutationListener(
+        GroupMentionNode,
+        (mutations, payload) => {
+          // Ignore mutations to GroupMentionNode (creation/updates/deletions) that are caused by collaboration (remote users) or history merge.
+          if (
+            payload.updateTags.has("collaboration") ||
+            payload.updateTags.has("history-merge")
+          ) {
+            return;
+          }
+
+          $handleGroupMentionMutation(mutations, payload);
+        }
+      );
+
+    return () => {
+      unsubscribeMentionMutationListener();
+      unsubscribeGroupMentionMutationListener();
+    };
   }, [editor, createTextMention, deleteTextMention]);
 
   useEffect(() => {
@@ -241,7 +315,9 @@ export function MentionPlugin() {
         if (nodes.length !== 1) return false;
 
         const node = nodes[0];
-        if (!$isMentionNode(node)) return false;
+        if (!$isMentionNode(node) && !$isGroupMentionNode(node)) {
+          return false;
+        }
 
         const text = $createTextNode("@");
         node.replace(text);
@@ -257,7 +333,10 @@ export function MentionPlugin() {
 
         const anchor = selection.anchor.getNode();
         const prevSibling = anchor.getPreviousSibling();
-        if (selection.anchor.offset === 0 && $isMentionNode(prevSibling)) {
+        if (
+          selection.anchor.offset === 0 &&
+          ($isMentionNode(prevSibling) || $isGroupMentionNode(prevSibling))
+        ) {
           const text = $createTextNode("@");
           prevSibling.replace(text);
 
@@ -269,7 +348,9 @@ export function MentionPlugin() {
           return true;
         } else if ($isElementNode(anchor)) {
           const child = anchor.getChildAtIndex(selection.anchor.offset - 1);
-          if (!$isMentionNode(child)) return false;
+          if (!$isMentionNode(child) && !$isGroupMentionNode(child)) {
+            return false;
+          }
 
           const text = $createTextNode("@");
           child.replace(text);
@@ -321,16 +402,19 @@ export function MentionPlugin() {
         const startOffset = selectionOffset - queryOffset;
         if (startOffset < 0) return;
 
-        let mentionNode: MentionNode;
+        let mentionNode: MentionNode | GroupMentionNode;
 
-        // Other mention kinds will be different nodes: GroupMentionNode, etc.
         switch (mention.kind) {
           case "user":
             mentionNode = $createMentionNode(mention.id);
             break;
 
+          case "group":
+            mentionNode = $createGroupMentionNode(mention.id, mention.userIds);
+            break;
+
           default:
-            return assertNever(mention.kind, "Unhandled mention kind");
+            return assertNever(mention, "Unhandled mention kind");
         }
 
         // Split the anchor (text) node and create a new text node only containing matched text.
@@ -369,14 +453,14 @@ export function MentionPlugin() {
           >
             <Suggestions.List className="lb-lexical-suggestions-list lb-lexical-mention-suggestions-list">
               {suggestions.map((mention) => {
-                switch (mention.kind) {
-                  case "user":
-                    return (
-                      <Suggestions.Item
-                        key={mention.id}
-                        value={mention.id}
-                        className="lb-lexical-suggestions-list-item lb-lexical-mention-suggestion"
-                      >
+                return (
+                  <Suggestions.Item
+                    key={mention.id}
+                    value={mention.id}
+                    className="lb-lexical-suggestions-list-item lb-lexical-mention-suggestion"
+                  >
+                    {mention.kind === "user" ? (
+                      <>
                         <Avatar
                           userId={mention.id}
                           className="lb-lexical-mention-suggestion-avatar"
@@ -385,12 +469,29 @@ export function MentionPlugin() {
                           userId={mention.id}
                           className="lb-lexical-mention-suggestion-user"
                         />
-                      </Suggestions.Item>
-                    );
-
-                  default:
-                    return assertNever(mention.kind, "Unhandled mention kind");
-                }
+                      </>
+                    ) : mention.kind === "group" ? (
+                      <>
+                        <Avatar
+                          groupId={mention.id}
+                          className="lb-lexical-mention-suggestion-avatar"
+                          icon={<UsersIcon />}
+                        />
+                        <Group
+                          groupId={mention.id}
+                          className="lb-lexical-mention-suggestion-group"
+                        >
+                          <GroupDescription
+                            groupId={mention.id}
+                            className="lb-lexical-mention-suggestion-group-description"
+                          />
+                        </Group>
+                      </>
+                    ) : (
+                      assertNever(mention, "Unhandled mention kind")
+                    )}
+                  </Suggestions.Item>
+                );
               })}
             </Suggestions.List>
           </SuggestionsPortal>