+++ packages/liveblocks-react/src/__tests__/useInboxNotifications.test.tsx
@@ -58,6 +58,7 @@ describe("useInboxNotifications", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -112,6 +113,7 @@ describe("useInboxNotifications", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -186,6 +188,7 @@ describe("useInboxNotifications", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -204,6 +207,7 @@ describe("useInboxNotifications", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -290,6 +294,7 @@ describe("useInboxNotifications", () => {
             threads: [],
             inboxNotifications: [],
             subscriptions: [],
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -460,6 +465,7 @@ describe("useInboxNotifications - Suspense", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -553,6 +559,7 @@ describe("useInboxNotifications: polling", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -765,6 +772,7 @@ describe("useInboxNotifications: polling", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -859,6 +867,7 @@ describe("useInboxNotifications: polling", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -1024,6 +1033,7 @@ describe("useInboxNotificationsSuspense: error", () => {
             threads,
             inboxNotifications,
             subscriptions,
+            groups: [],
             meta: {
               requestedAt: new Date().toISOString(),
               nextCursor: null,
@@ -1131,6 +1141,7 @@ describe("useInboxNotifications: pagination", () => {
               threads: [thread2],
               inboxNotifications: inboxNotificationsPage2,
               subscriptions: subscriptionsPage2,
+              groups: [],
               meta: {
                 requestedAt: new Date().toISOString(),
                 nextCursor: "cursor-2",
@@ -1146,6 +1157,7 @@ describe("useInboxNotifications: pagination", () => {
               threads: [thread3],
               inboxNotifications: inboxNotificationsPage3,
               subscriptions: subscriptionsPage3,
+              groups: [],
               meta: {
                 requestedAt: new Date().toISOString(),
                 nextCursor: "cursor-3",
@@ -1161,6 +1173,7 @@ describe("useInboxNotifications: pagination", () => {
               threads: [thread1],
               inboxNotifications: inboxNotificationsPage1,
               subscriptions: subscriptionsPage1,
+              groups: [],
               meta: {
                 requestedAt: new Date().toISOString(),
                 nextCursor: "cursor-1",
@@ -1306,6 +1319,7 @@ describe("useInboxNotifications: pagination", () => {
               threads: [threadTwo],
               inboxNotifications: inboxNotificationsPageTwo,
               subscriptions: subscriptionsPageTwo,
+              groups: [],
               meta: {
                 requestedAt: new Date().toISOString(),
                 nextCursor: null,
@@ -1320,6 +1334,7 @@ describe("useInboxNotifications: pagination", () => {
               threads: [threadOne],
               inboxNotifications: inboxNotificationsPageOne,
               subscriptions: subscriptionsPageOne,
+              groups: [],
               meta: {
                 requestedAt: new Date().toISOString(),
                 nextCursor: "cursor-1",
@@ -1408,6 +1423,7 @@ describe("useInboxNotifications: pagination", () => {
               threads: [threadOne],
               inboxNotifications: inboxNotificationsPageOne,
               subscriptions: subscriptionsPageOne,
+              groups: [],
               meta: {
                 requestedAt: new Date().toISOString(),
                 nextCursor: "cursor-1",