+++ packages/liveblocks-react/src/types/index.ts
@@ -28,7 +28,9 @@ import type {
   CommentAttachment,
   CommentBody,
   CommentData,
+  DGI,
   DRI,
+  GroupData,
   HistoryVersion,
   InboxNotificationData,
   LiveblocksError,
@@ -179,8 +181,14 @@ export type UserAsyncSuccess<T> = AsyncSuccess<T, "user">;
 export type RoomInfoAsyncResult = AsyncResult<DRI, "info">;
 export type RoomInfoAsyncSuccess = AsyncSuccess<DRI, "info">;
 
+export type GroupInfoAsyncResult = AsyncResult<DGI, "info">;
+export type GroupInfoAsyncSuccess = AsyncSuccess<DGI, "info">;
+
 export type AttachmentUrlAsyncResult = AsyncResult<string, "url">;
 export type AttachmentUrlAsyncSuccess = AsyncSuccess<string, "url">;
+
+export type GroupAsyncResult = AsyncResult<GroupData | undefined, "group">;
+export type GroupAsyncSuccess = AsyncSuccess<GroupData | undefined, "group">;
 
 // prettier-ignore
 export type CreateThreadOptions<M extends BaseMetadata> =
@@ -379,6 +387,14 @@ export type SharedContextBundle<U extends BaseUserMeta> = {
     useRoomInfo(roomId: string): RoomInfoAsyncResult;
 
     /**
+     * Returns group info from a given group ID.
+     *
+     * @example
+     * const { info, error, isLoading } = useGroupInfo("group-id");
+     */
+    useGroupInfo(groupId: string): GroupInfoAsyncResult;
+
+    /**
      * Returns whether the hook is called within a RoomProvider context.
      *
      * @example
@@ -454,6 +470,14 @@ export type SharedContextBundle<U extends BaseUserMeta> = {
      * const { info } = useRoomInfo("room-id");
      */
     useRoomInfo(roomId: string): RoomInfoAsyncSuccess;
+
+    /**
+     * Returns group info from a given group ID.
+     *
+     * @example
+     * const { info } = useGroupInfo("group-id");
+     */
+    useGroupInfo(groupId: string): GroupInfoAsyncSuccess;
 
     /**
      * Returns whether the hook is called within a RoomProvider context.